Global datatype declarations, syntax improvements
The syntax for inline datatypes was changed to use `$` and `;` instead
of `<` and `>`. Ex: `<T : Type | A : T>` becomes `$T : Type | A : T;`.

A global `data` syntax for datatypes was added. For example, this:

    data MyType : motive_type
    | ctor0     : ctor0_type
    | ctor1     : ctor1_type;

    <rest_of_code>

Is equivalent to:

    -- datatype declaration alias (uppercase)
    def MYTYPE
      $ MyType : motive_type
      | ctor0  : ctor0_type
      | ctor1  : ctor1_type;

    -- datatype declaration alias
    def MyType MYTYPE

    -- datatype Church-type alias
    def MyType &MYTYPE

    -- datatype constructor aliases
    def ctor0  @MYTYPE.ctor0
    def ctor1  @MYTYPE.ctor1

    <rest_of_code>

Moreover, the parser for constructor access and induction looks up terms
that normalize to datatype declarations, allowing the same variable name
"MyType" to be used for the type of the Church-encoding, and for the
datatype declaration itself. Those changes remove a lot of noise from
code, as seen on example.