Use dependent intersections
The good thing about dependent intersections is that they are much
simpler to implement, as can be seen by comparing the new version of
formality.js with the previous one. Moreover, being able to prove
induction for mendler encodings may be important in the long term. Also,
it is much more elegant and well principled.

My main complaint about this approach is that actually using it is
complex, as can be seen by comparing the new implementation of
bool_induction with the previous one. Now, instead of a straightforward
pattern analysis, we need 3 different implementations of booleans
(CBool, IBool, Bool), plus a proof of reflexivity. Programming like that
is obviously not practical. But now I realize there are many good ways
to solve this problem. Not only a language with inductive datatypes that
elaborates to it, but also something as simple as a preprocessor
language, coupled with libs for generics, could do it aswell.

At this point, thus, formality.js is just Cedille-Core. Some needed
changes such as EAL checks (for nasic compilation) could turn it into
something different, though. As such, I'm keeping the name for now. Yet,
depending on how things go, formality.js should possibly be renamed to
cedille.js.