`@T = a & b` syntax for dep.intersection value
Now, instead of `@ self : (B self) = a & b`, it is `@ T = a & b`, where
`T` is the dependent intersection type. This simplifies the code and
makes it more organized, because now we need only one declaration
(instead of two) to build a new inductive view on top of a Church-encoded
value, as demonstrated on the new `Ind` (inductive Nat) implementation.