New compilation cases, use unsafe for now
Since 1. code becomes significantly more complex with stratification
annotations and 2. I'm not sure if it is possible to prove nat_induction
on the stratified language, I'll be using the unsafe compilation to
inets (i.e., one that assigns unique ids to fan nodes, instead of using
the user-provided levels). This allows us to compile most "normal"
programs fine, but isn't guaranteed to work in all cases and, thus, not
safe to use for untrusted code. Meanwhile, we should research how to
work with the stratified annotations.