Replace dependent intersections by recursive self types
With dependent intersections, in order to implement an inductive
datatype, one must create first a simple view of that type (such as
CNat), then an inductive view (INat), and then intersect both views with
a dependent intersection, creating the actual type Nat. This is
necessary because inductive datatypes such as Nat must have access to
their constructors before they are defined.

If, though, we work in a language with mutual recursion, then this
problem doesn't exist. This allows us to replace dependent intersections
by self types. This has several benefits:

1. Inductive datatypes now involve only one type instead of 3!

2. Proving induction doesn't require proving reflexity. This means we
don't need equality primitives on the language.

3. We can work with Scott-encoded datatypes which have O(1) predecessor
and a very efficient representation on interaction combinators.

Not only the language implementation, but the proofs on it become much
simpler, as can be seen on formality.js and examples.js.

This comes with a huge cost, though: now the type checker procedure may
fail to halt.  This is because it is impossible to determine if two
non-normalizing λ-terms are equal. As such, this implementation tries
its best to identify equality between possibly unhalting terms, but, in
many cases, it will give up and let the user know it couldn't decide
equality.

Now, when it comes to consistency, I believe the issue is simpler. Due
to Curry-style, types aren't an essential part of terms, as they erase
to an underlying untyped language. That language is, usually, the pure
lambda calculus. If, though, we replace it by the EAL-calculus; i.e.,
the linear lambda calculus with elementary-affine logic duplications,
then it follows that our untyped language is terminating. This solves
two problems at once: 1. our language now is fully compatible with
interaction combinators; 2. it is terminating and thus consistent. This
will be added in a future commit.

As a last note, I noticed there seems to be an incompatibility between
Cedille-Core and EAL. The problem is that, as seen on earlier commits,
the term `to_ind : CNat -> Nat` can't be implemented on it. The best we
can do, I think, is `to_ind : CNat -> !Nat`, which can't be used to
implement induction. The reason is that the only way to implement
`to_ind` seems to be as `λn. (n succ zero)`, whereas what we'd really
need to preserve the level is the equivalent of `λn. λs. λz. (n s z)`.
As such, it seems to me that moving away from dependent intersections to
self types may be unavoidable.

// Proving induction on CoC

~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

// WITH SELF TYPES AND MUTUAL RECURSION
// Lambda (term):  [x : A] t
// Lambda (type):  {x : A} B
// Lambda (appl):  (f x)
// Self (type):    @ self = type
// Self (term):    : type = term
// Definition:     def x a b

// Natural numbers

def Nat @ self =
  {-Nat. : {self : Nat} Type}
  {succ. : {-pred : Nat} {_pred : (Nat. pred)} (Nat. (succ pred))}
  {zero. : (Nat. zero)}
  (Nat. self)

def succ [pred : Nat] : Nat =
  [-Nat. : {self : Nat} Type]
  [succ. : {-pred : Nat} {_pred : (Nat. pred)} (Nat. (succ pred))]
  [zero. : (Nat. zero)]
  (succ. -pred (~pred -Nat. succ. zero.))

def zero : Nat =
  [-Nat. : {self : Nat} Type]
  [succ. : {-pred : Nat} {_pred : (Nat. pred)} (Nat. (succ pred))]
  [zero. : (Nat. zero)]
  zero.

// Induction

def induction
  [self  : Nat]
  [-Nat. : {self : Nat} Type]
  [succ. : {-pred : Nat} {_pred : (Nat. pred)} (Nat. (succ pred))]
  [zero. : (Nat. zero)]
  (~self -Nat. succ. zero.)

~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

// WITH DEPENDENT INTERSECTIONS AND EQUALITY
// Lambda (term):        [x : A] t
// Lambda (type):        {x : A} B
// Lambda (appl):        (f x)
// Self (type):          @ self = type
// Self (term):          : type = term
// Definition:           def x a b
// Intersection (type):  <x : A> B
// Intersection (term):  @T = a & b
// Intersection (.fst):  .t
// Intersection (.snd):  +t
// Equality (type):      |a = b|
// Equality (refl):      $t k
// Equality (symm):      ~t
// Equality (cast):      ^e a b
// Equality (subs):      %x (P x) e t

// Non-inductive natural numbers

def Nat
    {-Nat : Type}
    {succ : {x : Nat} Nat}
    {zero : Nat}
    Nat

def succ
  [n    : Nat]
  [-Nat : Type]
  [succ : {x : Nat} Nat]
  [zero : Nat]
  (succ (n -Nat succ zero))

def zero
  [-Nat : Type]
  [succ : {x : Nat} Nat]
  [zero : Nat]
  zero

// Inductive natural numbers

def Ind <self : Nat>
  {-Ind : {n : Nat} Type}
  {step : {-n : Nat} {i : (Ind n)} (Ind (succ n))}
  {base : (Ind zero)}
  (Ind self)

def step [n : Ind] : Ind = (succ .n) &
  [-Ind : {n : Nat} Type]
  [step : {-n : Nat} {i : (Ind n)} (Ind (succ n))]
  [base : (Ind zero)]
  (step -.n (+n -Ind step base))

def base : Ind = zero &
  [-Ind : {n : Nat} Type]
  [step : {-n : Nat} {i : (Ind n)} (Ind (succ n))]
  [base : (Ind zero)]
  base

// Reflexivity

def to_ind [n : Nat]
  (n -Ind step base)

def ind_reflection [i : Ind]
  let motive [n : Nat]
    |.(to_ind n) = n|
  let case_succ [-n : Nat] [i : |.(to_ind n) = n|]
    (cong -Nat -Nat -.(to_ind n) -n -i -succ)
  let case_zero
    ($zero [x] x)
  (+i -motive case_succ case_zero)

// Induction

def induction
  [i  : Ind]
  [-P : {i : Ind} Type]
  [S  : {-i : Ind} {ih : (P i)} (P (step i))]
  [Z  : (P base)]
  let motive [n : Nat]
    (P (to_ind n))
  let case_succ [-n : Nat] [ih : (P (to_ind n))]
    (S -(to_ind n) ih)
  let case_zero
    Z
  (%i (P i) (ind_reflection i) (+i -motive case_succ case_zero))